pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[project]
name = "temporal-openai-agents-sdk"
version = "0.1.0"
description = "A repository for testing and learning Temporal and OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "temporalio>=1.5.1",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
